todo_service/requirements.txt
alembic==1.11.1
anyio==3.6.2
asyncpg==0.27.0
certifi==2023.5.7
click==8.1.3
distlib==0.3.6
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.2
filelock==3.12.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
orjson==3.8.14
pipenv==2023.5.19
platformdirs==3.5.1
psycopg2==2.9.6
pydantic==1.10.8
python-dotenv==1.0.0
python-multipart==0.0.6
pyjwt==2.6.0
passlib==1.7.4
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
typing_extensions==4.6.2
ujson==5.7.0
uvicorn==0.22.0
uvloop==0.17.0
virtualenv==20.23.0
virtualenv-clone==0.5.7
watchfiles==0.19.0
websockets==11.0.3